backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
pyyaml
scikit-learn
httpx